Refactoring code.
Seperating scanner and CLI tool into separate files. Not yet functional.
